+++ core/src/avm1/function.rs
@@ -578,22 +578,8 @@ impl<'gc> FunctionObject<'gc> {
 }
 
 impl<'gc> TObject<'gc> for FunctionObject<'gc> {
-    fn get_local_stored(
-        &self,
-        name: impl Into<AvmString<'gc>>,
-        activation: &mut Activation<'_, 'gc, '_>,
-    ) -> Option<Value<'gc>> {
-        self.base.get_local_stored(name, activation)
-    }
-
-    fn set_local(
-        &self,
-        name: AvmString<'gc>,
-        value: Value<'gc>,
-        activation: &mut Activation<'_, 'gc, '_>,
-        this: Object<'gc>,
-    ) -> Result<(), Error<'gc>> {
-        self.base.set_local(name, value, activation, this)
+    fn raw_script_object(&self) -> ScriptObject<'gc> {
+        self.base
     }
 
     fn call(
@@ -716,22 +702,6 @@ impl<'gc> TObject<'gc> for FunctionObject<'gc> {
         }
     }
 
-    fn getter(
-        &self,
-        name: AvmString<'gc>,
-        activation: &mut Activation<'_, 'gc, '_>,
-    ) -> Option<Object<'gc>> {
-        self.base.getter(name, activation)
-    }
-
-    fn setter(
-        &self,
-        name: AvmString<'gc>,
-        activation: &mut Activation<'_, 'gc, '_>,
-    ) -> Option<Object<'gc>> {
-        self.base.setter(name, activation)
-    }
-
     fn create_bare_object(
         &self,
         activation: &mut Activation<'_, 'gc, '_>,
@@ -750,167 +720,12 @@ impl<'gc> TObject<'gc> for FunctionObject<'gc> {
         .into())
     }
 
-    fn delete(&self, activation: &mut Activation<'_, 'gc, '_>, name: AvmString<'gc>) -> bool {
-        self.base.delete(activation, name)
-    }
-
-    fn proto(&self, activation: &mut Activation<'_, 'gc, '_>) -> Value<'gc> {
-        self.base.proto(activation)
-    }
-
-    fn define_value(
-        &self,
-        gc_context: MutationContext<'gc, '_>,
-        name: impl Into<AvmString<'gc>>,
-        value: Value<'gc>,
-        attributes: Attribute,
-    ) {
-        self.base.define_value(gc_context, name, value, attributes)
-    }
-
-    fn set_attributes(
-        &self,
-        gc_context: MutationContext<'gc, '_>,
-        name: Option<AvmString<'gc>>,
-        set_attributes: Attribute,
-        clear_attributes: Attribute,
-    ) {
-        self.base
-            .set_attributes(gc_context, name, set_attributes, clear_attributes)
-    }
-
-    fn add_property(
-        &self,
-        gc_context: MutationContext<'gc, '_>,
-        name: AvmString<'gc>,
-        get: Object<'gc>,
-        set: Option<Object<'gc>>,
-        attributes: Attribute,
-    ) {
-        self.base
-            .add_property(gc_context, name, get, set, attributes)
-    }
-
-    fn add_property_with_case(
-        &self,
-        activation: &mut Activation<'_, 'gc, '_>,
-        name: AvmString<'gc>,
-        get: Object<'gc>,
-        set: Option<Object<'gc>>,
-        attributes: Attribute,
-    ) {
-        self.base
-            .add_property_with_case(activation, name, get, set, attributes)
-    }
-
-    fn call_watcher(
-        &self,
-        activation: &mut Activation<'_, 'gc, '_>,
-        name: AvmString<'gc>,
-        value: &mut Value<'gc>,
-        this: Object<'gc>,
-    ) -> Result<(), Error<'gc>> {
-        self.base.call_watcher(activation, name, value, this)
-    }
-
-    fn watch(
-        &self,
-        activation: &mut Activation<'_, 'gc, '_>,
-        name: AvmString<'gc>,
-        callback: Object<'gc>,
-        user_data: Value<'gc>,
-    ) {
-        self.base.watch(activation, name, callback, user_data);
-    }
-
-    fn unwatch(&self, activation: &mut Activation<'_, 'gc, '_>, name: AvmString<'gc>) -> bool {
-        self.base.unwatch(activation, name)
-    }
-
-    fn has_property(&self, activation: &mut Activation<'_, 'gc, '_>, name: AvmString<'gc>) -> bool {
-        self.base.has_property(activation, name)
-    }
-
-    fn has_own_property(
-        &self,
-        activation: &mut Activation<'_, 'gc, '_>,
-        name: AvmString<'gc>,
-    ) -> bool {
-        self.base.has_own_property(activation, name)
-    }
-
-    fn has_own_virtual(
-        &self,
-        activation: &mut Activation<'_, 'gc, '_>,
-        name: AvmString<'gc>,
-    ) -> bool {
-        self.base.has_own_virtual(activation, name)
-    }
-
-    fn is_property_enumerable(
-        &self,
-        activation: &mut Activation<'_, 'gc, '_>,
-        name: AvmString<'gc>,
-    ) -> bool {
-        self.base.is_property_enumerable(activation, name)
-    }
-
-    fn get_keys(&self, activation: &mut Activation<'_, 'gc, '_>) -> Vec<AvmString<'gc>> {
-        self.base.get_keys(activation)
-    }
-
-    fn interfaces(&self) -> Vec<Object<'gc>> {
-        self.base.interfaces()
-    }
-
-    /// Set the interface list for this object. (Only useful for prototypes.)
-    fn set_interfaces(&self, gc_context: MutationContext<'gc, '_>, iface_list: Vec<Object<'gc>>) {
-        self.base.set_interfaces(gc_context, iface_list)
-    }
-
-    fn raw_script_object(&self) -> ScriptObject<'gc> {
-        self.base
-    }
-
     fn as_executable(&self) -> Option<Executable<'gc>> {
         self.data.read().function.clone()
     }
 
     fn as_ptr(&self) -> *const ObjectPtr {
         self.base.as_ptr()
-    }
-
-    fn length(&self, activation: &mut Activation<'_, 'gc, '_>) -> Result<i32, Error<'gc>> {
-        self.base.length(activation)
-    }
-
-    fn set_length(
-        &self,
-        activation: &mut Activation<'_, 'gc, '_>,
-        length: i32,
-    ) -> Result<(), Error<'gc>> {
-        self.base.set_length(activation, length)
-    }
-
-    fn has_element(&self, activation: &mut Activation<'_, 'gc, '_>, index: i32) -> bool {
-        self.base.has_element(activation, index)
-    }
-
-    fn get_element(&self, activation: &mut Activation<'_, 'gc, '_>, index: i32) -> Value<'gc> {
-        self.base.get_element(activation, index)
-    }
-
-    fn set_element(
-        &self,
-        activation: &mut Activation<'_, 'gc, '_>,
-        index: i32,
-        value: Value<'gc>,
-    ) -> Result<(), Error<'gc>> {
-        self.base.set_element(activation, index, value)
-    }
-
-    fn delete_element(&self, activation: &mut Activation<'_, 'gc, '_>, index: i32) -> bool {
-        self.base.delete_element(activation, index)
     }
 }
