+++ core/src/avm1/object.rs
@@ -100,7 +100,9 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         &self,
         name: impl Into<AvmString<'gc>>,
         activation: &mut Activation<'_, 'gc, '_>,
-    ) -> Option<Value<'gc>>;
+    ) -> Option<Value<'gc>> {
+        self.raw_script_object().get_local_stored(name, activation)
+    }
 
     /// Retrieve a named property from the object, or its prototype.
     fn get(
@@ -153,7 +155,10 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         value: Value<'gc>,
         activation: &mut Activation<'_, 'gc, '_>,
         this: Object<'gc>,
-    ) -> Result<(), Error<'gc>>;
+    ) -> Result<(), Error<'gc>> {
+        self.raw_script_object()
+            .set_local(name, value, activation, this)
+    }
 
     /// Set a named property on this object, or its prototype.
     fn set(
@@ -215,7 +220,9 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         activation: &mut Activation<'_, 'gc, '_>,
         this: Value<'gc>,
         args: &[Value<'gc>],
-    ) -> Result<Value<'gc>, Error<'gc>>;
+    ) -> Result<Value<'gc>, Error<'gc>> {
+        self.raw_script_object().call(name, activation, this, args)
+    }
 
     /// Construct the underlying object, if this is a valid constructor, and returns the result.
     /// Calling this on something other than a constructor will return a new Undefined object.
@@ -280,14 +287,18 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         &self,
         name: AvmString<'gc>,
         activation: &mut Activation<'_, 'gc, '_>,
-    ) -> Option<Object<'gc>>;
+    ) -> Option<Object<'gc>> {
+        self.raw_script_object().getter(name, activation)
+    }
 
     /// Retrive a setter defined on this object.
     fn setter(
         &self,
         name: AvmString<'gc>,
         activation: &mut Activation<'_, 'gc, '_>,
-    ) -> Option<Object<'gc>>;
+    ) -> Option<Object<'gc>> {
+        self.raw_script_object().setter(name, activation)
+    }
 
     /// Construct a host object of some kind and return its cell.
     ///
@@ -305,14 +316,18 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
     /// Delete a named property from the object.
     ///
     /// Returns false if the property cannot be deleted.
-    fn delete(&self, activation: &mut Activation<'_, 'gc, '_>, name: AvmString<'gc>) -> bool;
+    fn delete(&self, activation: &mut Activation<'_, 'gc, '_>, name: AvmString<'gc>) -> bool {
+        self.raw_script_object().delete(activation, name)
+    }
 
     /// Retrieve the `__proto__` of a given object.
     ///
     /// The proto is another object used to resolve methods across a class of
     /// multiple objects. It should also be accessible as `__proto__` from
     /// `get`.
-    fn proto(&self, activation: &mut Activation<'_, 'gc, '_>) -> Value<'gc>;
+    fn proto(&self, activation: &mut Activation<'_, 'gc, '_>) -> Value<'gc> {
+        self.raw_script_object().proto(activation)
+    }
 
     /// Define a value on an object.
     ///
@@ -331,7 +346,10 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         name: impl Into<AvmString<'gc>>,
         value: Value<'gc>,
         attributes: Attribute,
-    );
+    ) {
+        self.raw_script_object()
+            .define_value(gc_context, name, value, attributes)
+    }
 
     /// Set the attributes of a given property.
     ///
@@ -346,7 +364,10 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         name: Option<AvmString<'gc>>,
         set_attributes: Attribute,
         clear_attributes: Attribute,
-    );
+    ) {
+        self.raw_script_object()
+            .set_attributes(gc_context, name, set_attributes, clear_attributes)
+    }
 
     /// Define a virtual property onto a given object.
     ///
@@ -365,7 +386,10 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         get: Object<'gc>,
         set: Option<Object<'gc>>,
         attributes: Attribute,
-    );
+    ) {
+        self.raw_script_object()
+            .add_property(gc_context, name, get, set, attributes)
+    }
 
     /// Define a virtual property onto a given object.
     ///
@@ -384,7 +408,10 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         get: Object<'gc>,
         set: Option<Object<'gc>>,
         attributes: Attribute,
-    );
+    ) {
+        self.raw_script_object()
+            .add_property_with_case(activation, name, get, set, attributes)
+    }
 
     /// Calls the 'watcher' of a given property, if it exists.
     fn call_watcher(
@@ -393,7 +420,10 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         name: AvmString<'gc>,
         value: &mut Value<'gc>,
         this: Object<'gc>,
-    ) -> Result<(), Error<'gc>>;
+    ) -> Result<(), Error<'gc>> {
+        self.raw_script_object()
+            .call_watcher(activation, name, value, this)
+    }
 
     /// Set the 'watcher' of a given property.
     ///
@@ -404,16 +434,23 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         name: AvmString<'gc>,
         callback: Object<'gc>,
         user_data: Value<'gc>,
-    );
+    ) {
+        self.raw_script_object()
+            .watch(activation, name, callback, user_data)
+    }
 
     /// Removed any assigned 'watcher' from the given property.
     ///
     /// The return value will indicate if there was a watcher present before this method was
     /// called.
-    fn unwatch(&self, activation: &mut Activation<'_, 'gc, '_>, name: AvmString<'gc>) -> bool;
+    fn unwatch(&self, activation: &mut Activation<'_, 'gc, '_>, name: AvmString<'gc>) -> bool {
+        self.raw_script_object().unwatch(activation, name)
+    }
 
     /// Checks if the object has a given named property.
-    fn has_property(&self, activation: &mut Activation<'_, 'gc, '_>, name: AvmString<'gc>) -> bool;
+    fn has_property(&self, activation: &mut Activation<'_, 'gc, '_>, name: AvmString<'gc>) -> bool {
+        self.raw_script_object().has_property(activation, name)
+    }
 
     /// Checks if the object has a given named property on itself (and not,
     /// say, the object's prototype or superclass)
@@ -421,7 +458,9 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         &self,
         activation: &mut Activation<'_, 'gc, '_>,
         name: AvmString<'gc>,
-    ) -> bool;
+    ) -> bool {
+        self.raw_script_object().has_own_property(activation, name)
+    }
 
     /// Checks if the object has a given named property on itself that is
     /// virtual.
@@ -429,23 +468,35 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         &self,
         activation: &mut Activation<'_, 'gc, '_>,
         name: AvmString<'gc>,
-    ) -> bool;
+    ) -> bool {
+        self.raw_script_object().has_own_virtual(activation, name)
+    }
 
     /// Checks if a named property appears when enumerating the object.
     fn is_property_enumerable(
         &self,
         activation: &mut Activation<'_, 'gc, '_>,
         name: AvmString<'gc>,
-    ) -> bool;
+    ) -> bool {
+        self.raw_script_object()
+            .is_property_enumerable(activation, name)
+    }
 
     /// Enumerate the object.
-    fn get_keys(&self, activation: &mut Activation<'_, 'gc, '_>) -> Vec<AvmString<'gc>>;
+    fn get_keys(&self, activation: &mut Activation<'_, 'gc, '_>) -> Vec<AvmString<'gc>> {
+        self.raw_script_object().get_keys(activation)
+    }
 
     /// Enumerate all interfaces implemented by this object.
-    fn interfaces(&self) -> Vec<Object<'gc>>;
+    fn interfaces(&self) -> Vec<Object<'gc>> {
+        self.raw_script_object().interfaces()
+    }
 
     /// Set the interface list for this object. (Only useful for prototypes.)
-    fn set_interfaces(&self, gc_context: MutationContext<'gc, '_>, iface_list: Vec<Object<'gc>>);
+    fn set_interfaces(&self, gc_context: MutationContext<'gc, '_>, iface_list: Vec<Object<'gc>>) {
+        self.raw_script_object()
+            .set_interfaces(gc_context, iface_list)
+    }
 
     /// Determine if this object is an instance of a class.
     ///
@@ -617,20 +668,28 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
     }
 
     /// Gets the length of this object, as if it were an array.
-    fn length(&self, activation: &mut Activation<'_, 'gc, '_>) -> Result<i32, Error<'gc>>;
+    fn length(&self, activation: &mut Activation<'_, 'gc, '_>) -> Result<i32, Error<'gc>> {
+        self.raw_script_object().length(activation)
+    }
 
     /// Sets the length of this object, as if it were an array.
     fn set_length(
         &self,
         activation: &mut Activation<'_, 'gc, '_>,
         length: i32,
-    ) -> Result<(), Error<'gc>>;
+    ) -> Result<(), Error<'gc>> {
+        self.raw_script_object().set_length(activation, length)
+    }
 
     /// Checks if this object has an element.
-    fn has_element(&self, activation: &mut Activation<'_, 'gc, '_>, index: i32) -> bool;
+    fn has_element(&self, activation: &mut Activation<'_, 'gc, '_>, index: i32) -> bool {
+        self.raw_script_object().has_element(activation, index)
+    }
 
     /// Gets a property of this object, as if it were an array.
-    fn get_element(&self, activation: &mut Activation<'_, 'gc, '_>, index: i32) -> Value<'gc>;
+    fn get_element(&self, activation: &mut Activation<'_, 'gc, '_>, index: i32) -> Value<'gc> {
+        self.raw_script_object().get_element(activation, index)
+    }
 
     /// Sets a property of this object, as if it were an array.
     fn set_element(
@@ -638,10 +697,15 @@ pub trait TObject<'gc>: 'gc + Collect + Debug + Into<Object<'gc>> + Clone + Copy
         activation: &mut Activation<'_, 'gc, '_>,
         index: i32,
         value: Value<'gc>,
-    ) -> Result<(), Error<'gc>>;
+    ) -> Result<(), Error<'gc>> {
+        self.raw_script_object()
+            .set_element(activation, index, value)
+    }
 
     /// Deletes a property of this object as if it were an array.
-    fn delete_element(&self, activation: &mut Activation<'_, 'gc, '_>, index: i32) -> bool;
+    fn delete_element(&self, activation: &mut Activation<'_, 'gc, '_>, index: i32) -> bool {
+        self.raw_script_object().delete_element(activation, index)
+    }
 }
 
 pub enum ObjectPtr {}
